Harden router learning adaptation internals
