+++ src/semantic-router/pkg/extproc/router_learning_bandit.go
@@ -1,194 +1,14 @@
 package extproc
 
 import (
-	"context"
 	"fmt"
 	"math"
 	"sort"
 	"strings"
-	"sync"
-	"time"
 
 	"github.com/vllm-project/semantic-router/src/semantic-router/pkg/config"
 	"github.com/vllm-project/semantic-router/src/semantic-router/pkg/selection"
 )
-
-const (
-	routerLearningMethodBandit routerLearningMethod = "bandit"
-
-	banditReasonBaseBest        = "base_best"
-	banditReasonDecisionBypass  = "decision_bypass"
-	banditReasonIdentityMissing = "identity_missing"
-	banditReasonNoCandidates    = "no_candidates"
-	banditReasonScoreWin        = "score_win"
-	banditReasonStateMissing    = "state_missing"
-)
-
-type routerLearningRuntime struct {
-	mu              sync.Mutex
-	config          *config.RouterConfig
-	bandit          routerLearningBanditState
-	elo             routerLearningEloState
-	personalization routerLearningPersonalizationState
-}
-
-type routerLearningBanditState struct {
-	arms map[string]map[string]*routerLearningBanditArmState
-}
-
-type routerLearningBanditArmState struct {
-	Impressions   int
-	FeedbackCount int
-	RewardSum     float64
-	LastUpdated   time.Time
-}
-
-type routerLearningBanditScore struct {
-	model            string
-	quality          float64
-	cost             float64
-	latency          float64
-	meanReward       float64
-	exploration      float64
-	score            float64
-	impressions      int
-	feedbackCount    int
-	knownRewardState bool
-}
-
-type routerLearningBanditScoreDiagnostic struct {
-	Model         string  `json:"model"`
-	Score         float64 `json:"score"`
-	Quality       float64 `json:"quality"`
-	Cost          float64 `json:"cost"`
-	Latency       float64 `json:"latency"`
-	RewardMean    float64 `json:"reward_mean"`
-	Exploration   float64 `json:"exploration"`
-	Impressions   int     `json:"impressions"`
-	FeedbackCount int     `json:"feedback_count"`
-}
-
-func newRouterLearningRuntime(cfg *config.RouterConfig) *routerLearningRuntime {
-	return &routerLearningRuntime{
-		config: cfg,
-		bandit: routerLearningBanditState{
-			arms: map[string]map[string]*routerLearningBanditArmState{},
-		},
-		elo: routerLearningEloState{
-			ratings: map[string]map[string]*routerLearningEloRating{},
-		},
-		personalization: routerLearningPersonalizationState{
-			preferences: map[string]map[string]*routerLearningPersonalizationModelState{},
-		},
-	}
-}
-
-func (r *OpenAIRouter) routerLearningRuntimeState() *routerLearningRuntime {
-	if r == nil {
-		return nil
-	}
-	r.routerLearningMu.Lock()
-	defer r.routerLearningMu.Unlock()
-	if r.routerLearningRuntime == nil {
-		r.routerLearningRuntime = newRouterLearningRuntime(r.Config)
-	}
-	return r.routerLearningRuntime
-}
-
-func (rt *routerLearningRuntime) UpdateFeedback(_ context.Context, feedback *selection.Feedback) int {
-	if rt == nil || feedback == nil {
-		return 0
-	}
-	if err := selection.NormalizeFeedback(feedback); err != nil {
-		return 0
-	}
-	updated := 0
-	if rt.updateBanditFeedback(feedback) {
-		updated++
-	}
-	if rt.updateEloFeedback(feedback) {
-		updated++
-	}
-	if rt.updatePersonalizationFeedback(feedback) {
-		updated++
-	}
-	return updated
-}
-
-func (rt *routerLearningRuntime) updateBanditFeedback(feedback *selection.Feedback) bool {
-	if rt == nil || feedback == nil || strings.TrimSpace(feedback.WinnerModel) == "" {
-		return false
-	}
-	cfg, ok := rt.banditFeedbackConfig()
-	if !ok {
-		return false
-	}
-	stateKey, ok := learningStateKeyFromParts(
-		cfg.EffectiveScope(),
-		feedback.DecisionName,
-		feedback.SessionID,
-		feedback.ConversationID,
-	)
-	if !ok {
-		return false
-	}
-	reward := 1.0
-	if feedback.Confidence > 0 {
-		reward = math.Min(1, feedback.Confidence)
-	}
-	if feedback.Tie {
-		reward = 0.5
-	}
-	rt.recordBanditReward(stateKey, feedback.WinnerModel, reward)
-	if loser := strings.TrimSpace(feedback.LoserModel); loser != "" && !feedback.Tie {
-		rt.recordBanditReward(stateKey, loser, 0)
-	}
-	return true
-}
-
-func (rt *routerLearningRuntime) banditFeedbackConfig() (config.BanditLearningConfig, bool) {
-	if rt == nil || rt.config == nil || !rt.config.RouterLearning.Enabled {
-		return config.BanditLearningConfig{}, false
-	}
-	cfg := rt.config.RouterLearning.Adaptations.Bandit
-	return cfg, cfg.Enabled
-}
-
-func (rt *routerLearningRuntime) recordBanditImpression(stateKey string, model string) {
-	if rt == nil || strings.TrimSpace(stateKey) == "" || strings.TrimSpace(model) == "" {
-		return
-	}
-	rt.mu.Lock()
-	defer rt.mu.Unlock()
-	arm := rt.banditArm(stateKey, model)
-	arm.Impressions++
-	arm.LastUpdated = time.Now()
-}
-
-func (rt *routerLearningRuntime) recordBanditReward(stateKey string, model string, reward float64) {
-	if rt == nil || strings.TrimSpace(stateKey) == "" || strings.TrimSpace(model) == "" {
-		return
-	}
-	rt.mu.Lock()
-	defer rt.mu.Unlock()
-	arm := rt.banditArm(stateKey, model)
-	arm.FeedbackCount++
-	arm.RewardSum += clamp01(reward)
-	arm.LastUpdated = time.Now()
-}
-
-func (rt *routerLearningRuntime) banditArm(stateKey string, model string) *routerLearningBanditArmState {
-	if rt.bandit.arms == nil {
-		rt.bandit.arms = map[string]map[string]*routerLearningBanditArmState{}
-	}
-	if rt.bandit.arms[stateKey] == nil {
-		rt.bandit.arms[stateKey] = map[string]*routerLearningBanditArmState{}
-	}
-	if rt.bandit.arms[stateKey][model] == nil {
-		rt.bandit.arms[stateKey][model] = &routerLearningBanditArmState{}
-	}
-	return rt.bandit.arms[stateKey][model]
-}
 
 func (r *OpenAIRouter) applyBanditLearning(
 	input routerLearningInput,
@@ -324,34 +144,6 @@ func (r *OpenAIRouter) scoreBanditCandidates(
 		return scores[i].score > scores[j].score
 	})
 	return scores
-}
-
-func (rt *routerLearningRuntime) banditSnapshot(stateKey string, model string) routerLearningBanditArmState {
-	if rt != nil {
-		rt.mu.Lock()
-		defer rt.mu.Unlock()
-	}
-	if rt == nil || rt.bandit.arms == nil || rt.bandit.arms[stateKey] == nil || rt.bandit.arms[stateKey][model] == nil {
-		return routerLearningBanditArmState{}
-	}
-	return *rt.bandit.arms[stateKey][model]
-}
-
-func (rt *routerLearningRuntime) banditTotalImpressions(stateKey string) int {
-	if rt != nil {
-		rt.mu.Lock()
-		defer rt.mu.Unlock()
-	}
-	if rt == nil || rt.bandit.arms == nil || rt.bandit.arms[stateKey] == nil {
-		return 0
-	}
-	total := 0
-	for _, arm := range rt.bandit.arms[stateKey] {
-		if arm != nil {
-			total += arm.Impressions
-		}
-	}
-	return total
 }
 
 func banditExplorationBonus(cfg config.BanditLearningConfig, totalImpressions int, armImpressions int) float64 {
@@ -527,73 +319,6 @@ func banditAllScores(scores []routerLearningBanditScore) map[string]float64 {
 		result[score.model] = score.score
 	}
 	return result
-}
-
-func banditLearningPolicy(
-	cfg config.BanditLearningConfig,
-	mode string,
-	action routerLearningAction,
-	reason string,
-	stateKey string,
-	scores []routerLearningBanditScore,
-	winner routerLearningBanditScore,
-) routerLearningPolicy {
-	policy := newRouterLearningPolicy(routerLearningMethodBandit)
-	policy.Mode = mode
-	policy.Scope = cfg.EffectiveScope()
-	policy.Action = action
-	policy.Reason = reason
-	policy.Set("algorithm", cfg.EffectiveAlgorithm())
-	policy.Set("goals", normalizedLearningGoals(cfg.Goals))
-	if stateKey != "" {
-		policy.Set("state_key_hash", shortLearningIdentityHash(stateKey))
-	}
-	if winner.model != "" {
-		policy.Set("selected_model", winner.model)
-		policy.Set("selected_score", roundLearningFloat(winner.score))
-	}
-	if len(scores) > 0 {
-		policy.Set("scores", banditScoreDiagnostics(scores))
-	}
-	return policy
-}
-
-func banditScoreDiagnostics(scores []routerLearningBanditScore) []routerLearningBanditScoreDiagnostic {
-	result := make([]routerLearningBanditScoreDiagnostic, 0, len(scores))
-	for _, score := range scores {
-		result = append(result, routerLearningBanditScoreDiagnostic{
-			Model:         score.model,
-			Score:         roundLearningFloat(score.score),
-			Quality:       roundLearningFloat(score.quality),
-			Cost:          roundLearningFloat(score.cost),
-			Latency:       roundLearningFloat(score.latency),
-			RewardMean:    roundLearningFloat(score.meanReward),
-			Exploration:   roundLearningFloat(score.exploration),
-			Impressions:   score.impressions,
-			FeedbackCount: score.feedbackCount,
-		})
-	}
-	return result
-}
-
-func (r *OpenAIRouter) observeBanditSelection(
-	input routerLearningInput,
-	composed routerLearningComposition,
-) {
-	cfg, ok := r.banditLearningConfig(input.ctx)
-	if !ok || composed.selectedModelRef == nil {
-		return
-	}
-	stateKey, ok := banditStateKeyFromRequest(cfg.EffectiveScope(), input)
-	if !ok {
-		return
-	}
-	runtime := r.routerLearningRuntimeState()
-	runtime.recordBanditImpression(stateKey, composed.selectedModelRef.Model)
-}
-
-func banditStateKeyFromRequest(scope string, input routerLearningInput) (string, bool) {
-	return learningStateKeyFromRequest(scope, input)
 }
 
 func selectedModelRefByModel(selCtx *selection.SelectionContext, model string) *config.ModelRef {
