+++ src/semantic-router/pkg/extproc/router_learning_personalization.go
@@ -4,163 +4,10 @@ import (
 	"fmt"
 	"sort"
 	"strings"
-	"time"
 
 	"github.com/vllm-project/semantic-router/src/semantic-router/pkg/config"
 	"github.com/vllm-project/semantic-router/src/semantic-router/pkg/selection"
 )
-
-const (
-	routerLearningMethodPersonalization routerLearningMethod = "personalization"
-
-	personalizationReasonBaseBest        = "base_best"
-	personalizationReasonDecisionBypass  = "decision_bypass"
-	personalizationReasonIdentityMissing = "identity_missing"
-	personalizationReasonNoCandidates    = "no_candidates"
-	personalizationReasonPreferenceWin   = "preference_win"
-	personalizationReasonStateMissing    = "state_missing"
-)
-
-type routerLearningPersonalizationState struct {
-	preferences map[string]map[string]*routerLearningPersonalizationModelState
-}
-
-type routerLearningPersonalizationModelState struct {
-	Model        string
-	Positive     float64
-	Negative     float64
-	Interactions int
-	LastUpdated  time.Time
-}
-
-type routerLearningPersonalizationScore struct {
-	model        string
-	baseScore    float64
-	preference   float64
-	score        float64
-	positive     float64
-	negative     float64
-	interactions int
-	known        bool
-}
-
-type routerLearningPersonalizationScoreDiagnostic struct {
-	Model        string  `json:"model"`
-	Score        float64 `json:"score"`
-	BaseScore    float64 `json:"base_score"`
-	Preference   float64 `json:"preference"`
-	Positive     float64 `json:"positive"`
-	Negative     float64 `json:"negative"`
-	Interactions int     `json:"interactions"`
-}
-
-func (rt *routerLearningRuntime) updatePersonalizationFeedback(feedback *selection.Feedback) bool {
-	cfg, ok := rt.personalizationFeedbackConfig()
-	if !ok || feedback == nil || strings.TrimSpace(feedback.UserID) == "" {
-		return false
-	}
-	stateKey, ok := personalizationStateKeyFromFeedback(cfg.EffectiveScope(), feedback)
-	if !ok {
-		return false
-	}
-	rt.updatePersonalizationPreferences(stateKey, feedback)
-	return true
-}
-
-func (rt *routerLearningRuntime) personalizationFeedbackConfig() (config.PersonalizationLearningConfig, bool) {
-	if rt == nil || rt.config == nil || !rt.config.RouterLearning.Enabled {
-		return config.PersonalizationLearningConfig{}, false
-	}
-	cfg := rt.config.RouterLearning.Adaptations.Personalization
-	return cfg, cfg.Enabled
-}
-
-func personalizationStateKeyFromFeedback(scope string, feedback *selection.Feedback) (string, bool) {
-	baseKey, ok := learningStateKeyFromParts(
-		scope,
-		feedback.DecisionName,
-		feedback.SessionID,
-		feedback.ConversationID,
-	)
-	if !ok {
-		return "", false
-	}
-	userID := strings.TrimSpace(feedback.UserID)
-	if userID == "" {
-		return "", false
-	}
-	return "user:" + userID + "/" + baseKey, true
-}
-
-func personalizationStateKeyFromRequest(scope string, input routerLearningInput) (string, bool) {
-	baseKey, ok := learningStateKeyFromRequest(scope, input)
-	if !ok {
-		return "", false
-	}
-	userID := learningUserIDFromRequest(input)
-	if userID == "" {
-		return "", false
-	}
-	return "user:" + userID + "/" + baseKey, true
-}
-
-func (rt *routerLearningRuntime) updatePersonalizationPreferences(stateKey string, feedback *selection.Feedback) {
-	if rt == nil || strings.TrimSpace(stateKey) == "" || feedback == nil {
-		return
-	}
-	rt.mu.Lock()
-	defer rt.mu.Unlock()
-
-	if feedback.Tie {
-		if feedback.WinnerModel != "" {
-			pref := rt.personalizationPreference(stateKey, feedback.WinnerModel)
-			pref.Positive += 0.5
-			pref.Interactions++
-			pref.LastUpdated = time.Now()
-		}
-		if feedback.LoserModel != "" {
-			pref := rt.personalizationPreference(stateKey, feedback.LoserModel)
-			pref.Positive += 0.5
-			pref.Interactions++
-			pref.LastUpdated = time.Now()
-		}
-		return
-	}
-
-	reward := 1.0
-	if feedback.Confidence > 0 {
-		reward = clamp01(feedback.Confidence)
-	}
-	if feedback.WinnerModel != "" {
-		pref := rt.personalizationPreference(stateKey, feedback.WinnerModel)
-		pref.Positive += reward
-		pref.Interactions++
-		pref.LastUpdated = time.Now()
-	}
-	if feedback.LoserModel != "" {
-		pref := rt.personalizationPreference(stateKey, feedback.LoserModel)
-		pref.Negative += reward
-		pref.Interactions++
-		pref.LastUpdated = time.Now()
-	}
-}
-
-func (rt *routerLearningRuntime) personalizationPreference(
-	stateKey string,
-	model string,
-) *routerLearningPersonalizationModelState {
-	if rt.personalization.preferences == nil {
-		rt.personalization.preferences = map[string]map[string]*routerLearningPersonalizationModelState{}
-	}
-	if rt.personalization.preferences[stateKey] == nil {
-		rt.personalization.preferences[stateKey] = map[string]*routerLearningPersonalizationModelState{}
-	}
-	model = strings.TrimSpace(model)
-	if rt.personalization.preferences[stateKey][model] == nil {
-		rt.personalization.preferences[stateKey][model] = &routerLearningPersonalizationModelState{Model: model}
-	}
-	return rt.personalization.preferences[stateKey][model]
-}
 
 func (r *OpenAIRouter) applyPersonalizationLearning(
 	input routerLearningInput,
@@ -426,54 +273,6 @@ func personalizationSelectionResult(
 		result.LoRAName = baseResult.LoRAName
 		result.Method = baseResult.Method
 		result.Tier = baseResult.Tier
-	}
-	return result
-}
-
-func personalizationLearningPolicy(
-	cfg config.PersonalizationLearningConfig,
-	mode string,
-	action routerLearningAction,
-	reason string,
-	stateKey string,
-	userID string,
-	scores []routerLearningPersonalizationScore,
-	winner routerLearningPersonalizationScore,
-) routerLearningPolicy {
-	policy := newRouterLearningPolicy(routerLearningMethodPersonalization)
-	policy.Mode = mode
-	policy.Scope = cfg.EffectiveScope()
-	policy.Action = action
-	policy.Reason = reason
-	if stateKey != "" {
-		policy.Set("state_key_hash", shortLearningIdentityHash(stateKey))
-	}
-	if userID != "" {
-		policy.Set("user_hash", shortLearningIdentityHash(userID))
-	}
-	if winner.model != "" {
-		policy.Set("selected_model", winner.model)
-		policy.Set("selected_score", roundLearningFloat(winner.score))
-		policy.Set("selected_preference", roundLearningFloat(winner.preference))
-	}
-	if len(scores) > 0 {
-		policy.Set("preferences", personalizationScoreDiagnostics(scores))
-	}
-	return policy
-}
-
-func personalizationScoreDiagnostics(scores []routerLearningPersonalizationScore) []routerLearningPersonalizationScoreDiagnostic {
-	result := make([]routerLearningPersonalizationScoreDiagnostic, 0, len(scores))
-	for _, score := range scores {
-		result = append(result, routerLearningPersonalizationScoreDiagnostic{
-			Model:        score.model,
-			Score:        roundLearningFloat(score.score),
-			BaseScore:    roundLearningFloat(score.baseScore),
-			Preference:   roundLearningFloat(score.preference),
-			Positive:     roundLearningFloat(score.positive),
-			Negative:     roundLearningFloat(score.negative),
-			Interactions: score.interactions,
-		})
 	}
 	return result
 }
